Human-Segmentor/requirements.txt
pytz==2021.1
torch==1.9.0
typing-extensions==3.10.0.2
Pillow==8.3.2
numpy==1.21.2
tqdm==4.62.2
torchvision==0.10.0
opencv-python==4.5.3.56
matplotlib==3.4.3
pytest==6.2.5
pandas==1.3.3
ipython==7.28.0
ipykernel==6.4.1
ray==1.8.0
bayesian-optimization==1.2.0
GPUtil==1.4.0